+++ src/vm_error.rs
@@ -1,0 +1,8 @@
+#[derive(Copy, Clone, Debug)]
+pub enum VMError {
+    InvalidOpCode(u8),
+    InvalidTrapCode(u8),
+    TerminalIOAttributesGet,
+    TerminalIOAttributesSet,
+    IOError,
+}
